+++ lib/utils/consts.dart
@@ -2,7 +2,6 @@ const privateFlag = 'private';
 const enableNotificationsFlag = 'enable_notifications';
 const notifyExceptionsFlag = 'notify_exceptions';
 const hideForScreenshotFlag = 'hide_for_screenshot';
-const showTasksTabFlag = 'show_tasks_tab';
 const showBrightSchemeFlag = 'show_bright_scheme';
 const showThemeConfigFlag = 'show_theme_config';
 const allowInvalidCertFlag = 'allow_invalid_cert';