+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed:
 - Remove redundant config flag: enable_beamer_nav
 - Remove redundant config flag: listen_to_global_screenshot_hotkey
+- Remove redundant config flag: show_tasks_tab
 
 ## [0.8.249] - 2023-01-23
 ### Changed: